+++ __fixtures__/!general.js
@@ -15,23 +15,6 @@ const MediaColorProperty = tw.div`lg:text-red-500`
 const ElementMediaColorProperty = tw(Box)`lg:bg-blue-500`
 const MediaPropertyDuplicates = tw`lg:bg-blue-500 lg:bg-black`
 
-// Important
-
-const Important = tw`lg:uppercase!`
-const MediaImportant = tw.div`lg:text-red-500!`
-const ElementMediaImportant = tw(Box)`lg:bg-blue-500!`
-
-const PlaceholderImportant = tw.input`placeholder-red-500!`
-const StateImportant = tw.input`hover:text-red-500!`
-const StatePlaceholderImportant = tw.input`hover:placeholder-red-500!`
-const StateStatePlaceholderImportant = tw.input`active:hover:placeholder-red-500!`
-const StateMultiplePropertiesImportant = tw.input`hover:truncate!`
-const MediaStateMultiplePropertiesImportant = tw.input`lg:hover:truncate!`
-const ElementMediaStateMultiplePropertiesImportant = tw(Box)`lg:hover:truncate!`
-
-const JsxPlaceholder = () => <input tw="placeholder-red-500" />
-const JsxPlaceholderImportant = () => <input tw="placeholder-green-500!" />
-
 // Only basic evaluations supported
 // No functions or "beyond basic" conditionals.
 const plainConditional = true && 'red'
